Preserve mtime when creating the non-digested file
FileUtils.cp does not copy the file-attributes from the source to the new
file.  With copy_file, this is possible. The third argument needs to be
"trueish", so a descriptive symbol suffices.

In effect, this copies over the mtime from the assets. This helps Webservers
to generate ETags based on the mtime which in turn can prevent the need to
retransfer a file if it hasn't changed since the last request.

Essentially, this patch is saving the money of all those who are paying for
bandwith. As a side-effect, I'm not confused when I see the directory
listing.